LayoutTests/media/remove-from-document-no-load-expected.txt

Test that removing a media element from the tree when no media has been loaded does not generate a loadstart event.

EXPECTED (video.networkState == '0') OK
RUN(document.body.removeChild(video))
EXPECTED (didReceiveLoadStartEvent == 'false') OK
END OF TEST
